Support overriding telemetry endpoint
